requirements/requirements.txt

Django==1.2.1
South==0.7.1
MySQL-python==1.2.3
PIL==1.1.7
flup

django-compressor==0.5.3
django-endless-pagination==0.6
django-notification==0.1.5
-e git+http://github.com/ericflo/django-avatar.git/@b52c89bbdd39aac4e5abc9e716ef0b2459014858#egg=django_avatar-1.1a5-py2.6-dev

# For Admin Interface
fullhistory==0.3.0-final
feedparser==4.1

 -e svn+http://django-grappelli.googlecode.com/svn/trunk@1093#egg=django_grappelli-2.0-py2.6-dev_r1093
-e hg+http://bitbucket.org/fetzig/grappelli-admin-tools/@4583eb7c9163d62fd4779eb098745a82a72e3ec8#egg=django_admin_tools-tip

# For openid
-e git+http://github.com/adieu/python-openid.git/@1423472e9a2f0d2ee8aa594a64dcd4824a266448#egg=python_openid-2.2.4-py2.6-dev
-e hg+http://bitbucket.org/dkmurthy/django-openid-auth@38ab365490878c0d0b3b50713c8cfe1b98e4f822#egg=django_openid_auth-tip

# For Google Apps
gdata==2.0.10

# For Developers
ipython==0.10
docutils==0.7
django-debug-toolbar==0.8.3
Fabric==0.9.1
pylint==0.21.1
nose==0.11.1
djangosanetesting==0.5.6
django-extensions==0.5
Werkzeug==0.6.2

## The following requirements were added by pip --freeze:
distribute
slimmer==0.1.30
logilab-astng==0.20.1
logilab-common==0.50.3
pycrypto==2.2
wsgiref==0.1.2
BeautifulSoup==3.1.0.1
